+++ micropip/freeze.py
@@ -1,10 +1,10 @@
 import json
 from copy import deepcopy
-from importlib.metadata import distributions
 from typing import Any
 
 from packaging.utils import canonicalize_name
 
+from . import _utils
 from ._compat import REPODATA_INFO, REPODATA_PACKAGES
 
 
@@ -22,7 +22,7 @@ def freeze() -> str:
     """
 
     packages = deepcopy(REPODATA_PACKAGES)
-    for dist in distributions():
+    for dist in _utils.importlib_distributions():
         name = dist.name
         version = dist.version
         url = dist.read_text("PYODIDE_URL")
