+++ tests/test_freeze.py
@@ -1,0 +1,32 @@
+import pytest
+from conftest import mock_fetch_cls
+
+
+@pytest.mark.asyncio
+async def test_freeze(mock_fetch: mock_fetch_cls) -> None:
+    import micropip
+
+    dummy = "dummy"
+    dep1 = "dep1"
+    dep2 = "dep2"
+    toplevel = [["abc", "def", "geh"], ["c", "h", "i"], ["a12", "b13"]]
+
+    mock_fetch.add_pkg_version(dummy, requirements=[dep1, dep2], top_level=toplevel[0])
+    mock_fetch.add_pkg_version(dep1, top_level=toplevel[1])
+    mock_fetch.add_pkg_version(dep2, top_level=toplevel[2])
+
+    await micropip.install(dummy)
+
+    import json
+
+    lockfile = json.loads(micropip.freeze())
+
+    pkg_metadata = lockfile["packages"][dummy]
+    dep1_metadata = lockfile["packages"][dep1]
+    dep2_metadata = lockfile["packages"][dep2]
+    assert pkg_metadata["depends"] == [dep1, dep2]
+    assert dep1_metadata["depends"] == []
+    assert dep2_metadata["depends"] == []
+    assert pkg_metadata["imports"] == toplevel[0]
+    assert dep1_metadata["imports"] == toplevel[1]
+    assert dep2_metadata["imports"] == toplevel[2]
